# docs/openapi/restaurants.yaml
openapi: 3.0.3
info:
  title: plazoleta-service API
  version: 1.0.0
  description: API for restaurants and plates
servers:
  - url: http://localhost:8084
security:
  - bearerAuth: []
paths:
  /api/v1/restaurants:
    get:
      tags: [Restaurants]
      summary: List restaurants (public)
      description: Returns restaurants ordered alphabetically by name. Only fields name and logoUrl are returned.
      security: []
      parameters:
        - in: query
          name: page
          required: false
          schema: { type: integer, default: 0, minimum: 0 }
        - in: query
          name: size
          required: false
          schema: { type: integer, default: 10, minimum: 1 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestaurantListItem'
    post:
      tags: [Restaurants]
      summary: Create restaurant (ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantCreateRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: URI of the created resource
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401': { description: Unauthorized }
        '403': { description: Forbidden (requires ADMIN) }
        '404':
          description: Owner not found or not OWNER role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: NIT already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/plates:
    post:
      tags: [Plates]
      summary: Create plate (OWNER of the restaurant)
      description: Only the restaurant owner can create plates for their restaurant. Initial status is active = true.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlateCreateRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: URI of the created resource
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlateResponse'
        '400':
          description: Validation error (price > 0, required fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401': { description: Unauthorized }
        '403':
          description: Forbidden (requires OWNER and ownership of the restaurant)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Plate with same name already exists in restaurant (if applicable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/plates/{id}:
    put:
      tags: [Plates]
      summary: Update plate price and description (OWNER of the restaurant)
      description: Only the restaurant owner can update plates of their restaurant. Only price and description are allowed to be modified.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlateUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlateResponse'
        '400':
          description: Validation error (price > 0, description required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401': { description: Unauthorized }
        '403':
          description: Forbidden (requires OWNER and ownership of the restaurant)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Plate not found or Restaurant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/plates/{id}/status:
    patch:
      tags: [Plates]
      summary: Enable/Disable plate (OWNER of the restaurant)
      description: Only the restaurant owner can enable/disable plates of their restaurant.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlateStatusUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401': { description: Unauthorized }
        '403':
          description: Forbidden (requires OWNER and ownership of the restaurant)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Plate not found or Restaurant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/plates/restaurant/{id}:
    get:
      tags: [Plates]
      summary: List active plates by restaurant (public)
      description: Returns active plates for a restaurant, optionally filtered by category, ordered by name.
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
        - in: query
          name: category
          required: false
          schema: { $ref: '#/components/schemas/PlateCategory' }
        - in: query
          name: page
          required: false
          schema: { type: integer, default: 0, minimum: 0 }
        - in: query
          name: size
          required: false
          schema: { type: integer, default: 10, minimum: 1 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlateResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RestaurantCreateRequest:
      type: object
      required: [name, nit, address, phone, logoUrl, ownerId]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          pattern: "^(?!\\d+$).+$"
        nit:
          type: string
          minLength: 1
          maxLength: 20
          pattern: "^\\d+$"
        address:
          type: string
          maxLength: 200
        phone:
          type: string
          maxLength: 13
          pattern: "^\\+?\\d{1,13}$"
        logoUrl:
          type: string
          format: uri
          maxLength: 255
        ownerId:
          type: integer
          format: int64
    RestaurantResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        nit: { type: string }
        address: { type: string }
        phone: { type: string }
        logoUrl: { type: string }
        ownerId: { type: integer, format: int64 }

    RestaurantListItem:
      type: object
      properties:
        name: { type: string }
        logoUrl: { type: string }

    PlateCreateRequest:
      type: object
      required: [name, price, description, imageUrl, category, restaurantId]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        price:
          type: integer
          minimum: 1
        description:
          type: string
          minLength: 1
          maxLength: 1000
        imageUrl:
          type: string
          format: uri
          maxLength: 255
        category:
          $ref: '#/components/schemas/PlateCategory'
        restaurantId:
          type: integer
          format: int64

    PlateResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        price: { type: integer }
        description: { type: string }
        imageUrl: { type: string }
        category: { $ref: '#/components/schemas/PlateCategory' }
        active: { type: boolean }
        restaurantId: { type: integer, format: int64 }

    PlateUpdateRequest:
      type: object
      required: [price, description]
      properties:
        price:
          type: integer
          minimum: 1
        description:
          type: string
          minLength: 1
          maxLength: 1000

    PlateStatusUpdateRequest:
      type: object
      required: [active]
      properties:
        active:
          type: boolean

    PlateCategory:
      type: string
      enum: [ENTRADA, PRINCIPAL, POSTRE, BEBIDA]

    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        path: { type: string }
        message: { type: string }
        code: { type: string }